handle v2 path
