+++ src/meta/src/stream/stream_manager.rs
@@ -867,6 +867,83 @@ impl GlobalStreamManager {
             table_fragment_map.into(),
         ))
     }
+
+    pub(crate) async fn recover_background_mv_progress_v2(
+        &self,
+    ) -> MetaResult<(
+        TableActorMap,
+        TableUpstreamMvCountMap,
+        TableDefinitionMap,
+        TableNotifierMap,
+        TableFragmentMap,
+    )> {
+        let mgr = self.metadata_manager.as_v2_ref();
+        let mviews = mgr
+            .catalog_controller
+            .list_background_creating_mviews()
+            .await?;
+
+        let mut senders = HashMap::new();
+        let mut receivers = Vec::new();
+        let mut table_fragment_map = HashMap::new();
+        let mut mview_definitions = HashMap::new();
+        let mut table_map = HashMap::new();
+        let mut upstream_mv_counts = HashMap::new();
+        for mview in &mviews {
+            let (finished_tx, finished_rx) = oneshot::channel();
+            let table_id = TableId::new(mview.table_id as _);
+            senders.insert(
+                table_id,
+                Notifier {
+                    finished: Some(finished_tx),
+                    ..Default::default()
+                },
+            );
+
+            let table_fragments = mgr
+                .catalog_controller
+                .get_job_fragments_by_id(mview.table_id)
+                .await?;
+            let table_fragments = TableFragments::from_protobuf(table_fragments);
+            upstream_mv_counts.insert(table_id, table_fragments.dependent_table_ids());
+            table_map.insert(table_id, table_fragments.backfill_actor_ids());
+            table_fragment_map.insert(table_id, table_fragments);
+            mview_definitions.insert(table_id, mview.definition.clone());
+            receivers.push((mview.table_id, finished_rx));
+        }
+
+        for (id, finished) in receivers {
+            let catalog_controller = mgr.catalog_controller.clone();
+            tokio::spawn(async move {
+                let res: MetaResult<()> = try {
+                    tracing::debug!("recovering stream job {}", id);
+                    finished.await.ok().context("failed to finish command")??;
+                    tracing::debug!(id, "finished stream job");
+                    catalog_controller.finish_streaming_job(id, None).await?;
+                };
+                if let Err(e) = &res {
+                    tracing::error!(
+                        id,
+                        error = %e.as_report(),
+                        "stream job interrupted, will retry after recovery",
+                    );
+                    // NOTE(kwannoel): We should not cleanup stream jobs,
+                    // we don't know if it's just due to CN killed,
+                    // or the job has actually failed.
+                    // Users have to manually cancel the stream jobs,
+                    // if they want to clean it.
+                }
+            });
+        }
+
+        Ok((
+            table_map.into(),
+            upstream_mv_counts.into(),
+            mview_definitions.into(),
+            senders.into(),
+            table_fragment_map.into(),
+        ))
+    }
 }
 
 #[cfg(test)]
